+++ bigml/constants.py
@@ -253,3 +253,31 @@ STORAGE = "./storage"
 
 # label for transient HTTP errors
 TRANSIENT = "transient"
+
+# fields related attributes
+RESOURCES_WITH_FIELDS = [SOURCE_PATH, DATASET_PATH, MODEL_PATH,
+                         PREDICTION_PATH, CLUSTER_PATH, ANOMALY_PATH,
+                         SAMPLE_PATH, CORRELATION_PATH, STATISTICAL_TEST_PATH,
+                         LOGISTIC_REGRESSION_PATH, ASSOCIATION_PATH,
+                         TOPIC_MODEL_PATH, ENSEMBLE_PATH, PCA_PATH,
+                         DEEPNET_PATH, LINEAR_REGRESSION_PATH]
+DEFAULT_MISSING_TOKENS = ["", "N/A", "n/a", "NULL", "null", "-", "#DIV/0",
+                          "#REF!", "#NAME?", "NIL", "nil", "NA", "na",
+                          "#VALUE!", "#NULL!", "NaN", "#N/A", "#NUM!", "?"]
+FIELDS_PARENT = { \
+    "model": "model",
+    "anomaly": "model",
+    "cluster": "clusters",
+    "logisticregression": "logistic_regression",
+    "linearregression": "linear_regression",
+    "ensemble": "ensemble",
+    "deepnet": "deepnet",
+    "topicmodel": "topic_model",
+    "association": "associations",
+    "fusion": "fusion",
+    "correlation": "correlations",
+    "sample": "sample",
+    "pca": "pca",
+    "timeseries": "timeseries",
+    "statisticaltest": "statistical_tests"}
+ALL_FIELDS = "limit=-1"